Create jQueryTutorialsL10.css
This css file, and others in this series, were not explained until this lesson -- and then incompletely.

I made changes in some of the other lessons to get the look of his example, and the changes he made in this video are also included. This final css file could be used for all lessons.